# public/swaggerOutput.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Meu Primeiro Alô",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "172.25.176.1:8102",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/secretary/referral-services/": {
      "get": {
        "tags": [
          "ReferralService"
        ],
        "description": "Endpoint para recuperar todos os serviços de referencia",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "ReferralService"
        ],
        "description": "Endpoint para adicionar um serviço de referencia",
        "parameters": [
          {
            "name": "referralService",
            "in": "body",
            "required": "true",
            "description": "Nome e telefone do contado",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "João S. da Silva"
                },
                "cellphone": {
                  "type": "string",
                  "example": "554130306905"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/secretary/referral-services/{id}": {
      "get": {
        "tags": [
          "ReferralService"
        ],
        "description": "Endpoint para recuperar um serviço de referencia pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "ReferralService"
        ],
        "description": "Endpoint para atualizar um serviço de referencia",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "referralService",
            "in": "body",
            "required": "true",
            "description": "Nome e telefone do contado",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "João S. da Silva"
                },
                "cellphone": {
                  "type": "string",
                  "example": "554130306905"
                }
              }
            }
          }
        ],
        "responses": {},
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "ReferralService"
        ],
        "description": "Endpoint para deletar um serviço de referencia pelo id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "definitions": {}
}